app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
pydantic>=2.7.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.29
asyncpg>=0.29.0
anthropic>=0.28.0
qdrant-client>=1.13.0,<1.15.0
sentence-transformers>=3.0.0
torch>=2.2.2
pymupdf>=1.24.0
pymupdf4llm>=0.0.5
langchain-text-splitters>=0.2.0
unstructured[docx,pptx,xlsx]>=0.13.0
boto3>=1.34.0
structlog>=24.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
# ── OpenTelemetry ─────────────────────────────────────────────────────────────
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
email-validator>=2.1.0
# ── Agentic tools ─────────────────────────────────────────────────────────────
yfinance>=0.2.37
pandas>=2.0.0
